Refactor ops/webauthn a bit to make it easier to handle
